requirements.txt
cassandra_driver==3.29.2
cassio==0.1.10
langchain==0.3.8
langchain_community==0.3.8
langchain_core==0.3.21
langchain_ollama==0.2.0
langchain_openai==0.2.9
langchain_text_splitters==0.3.2
langgraph==0.2.53
llama_index==0.12.1
pydantic==2.10.1
tqdm==4.67.0
typing_extensions==4.12.2
